Code refactor 2 electric boogaloo

+++ Assets/Scripts/ThirdPersonShooting.cs
@@ -6,7 +6,7 @@ using TMPro;
 
 public class ThirdPersonShooting : MonoBehaviour
 {
-    [SerializeField] int bulletDamage;
+    [SerializeField] float bulletDamage;
     public Camera cam;
 
     public GameObject projectile;
